requirements.txt
beautifulsoup4==4.11.1
discord.py==2.0.1
praw==7.6.0
python-dotenv==0.21.0
requests==2.27.1
stability_sdk==0.1.1
youtube_dl==2021.12.17
